==> 2.From_PDF_Detach_XML.ipynb <==
# instalar dependências
# !pip install subprocess
# pdfdetach.exe deve estar na pasta do Jupyter Notebook

import os
from subprocess import call

os.chdir('d:\\Users\\Patri\\Documents\\1.TCC\\Notebooks\\WebScrapAcordaosPDF')

pasta = os.listdir()
for count, arquivo in enumerate(pasta) : print (count, arquivo)

# pasta_qual = input ('Qual o número da pasta que deseja usar: ')
# os.chdir(pasta[int(pasta_qual)])
# os.getcwd()

os.listdir()

lista = [f for f in os.listdir('.') if os.path.isfile(f)]
len(lista)

# Extrair o arquivo XML de cada PDF usando pdfdetach
from datetime import datetime
if not 'XML' in os.listdir(): os.mkdir("XML")

inicio = datetime.now()
hora_inicio = inicio.strftime("%H:%M:%S")
inicio_string = inicio.strftime("%Y%m%B")

for count, arquivo in enumerate(lista):
    arquivo_nome_ext=lista[count]
    arquivo_nome=arquivo_nome_ext[:-4]
    arquivo_nome_xml=arquivo_nome+".xml"
    call(["pdfdetach", "-saveall", arquivo_nome_ext])
    if "documento.xml" in os.listdir() : os.rename("documento.xml", os.getcwd()+"\\XML\\"+arquivo_nome_xml)
    decorrido = datetime.now() - inicio
    expectativa = decorrido / (count+1) * len(lista)
    if decorrido > expectativa: 
        restante = 0 
    else: 
        restante = expectativa - decorrido
    print (count, str(decorrido)[:-7], str(expectativa)[:-7], str(restante)[:-7], arquivo_nome_xml)

len(lista)

len(os.listdir())

os.getcwd()
