==> 3.Select_Versao_PGD.ipynb <==
import xml.etree.ElementTree as et
import pandas as pd
import numpy as np
import os
import re
import pprint

# os.chdir('d:\\Users\\Patri\\Documents\\1.TCC\\Notebooks\\WebScrapAcordaosPDF')

pasta = os.listdir()
for count, arquivo in enumerate(pasta) : print (count, arquivo)

# pasta_qual = input ('Qual o número da pasta que deseja usar: ')
# os.chdir(pasta[int(pasta_qual)]+"\\XML")
# os.getcwd()

os.listdir()

# Separar arquivos XML por diretório: Resolução x Acórdão/versão PGD

from datetime import datetime

inicio = datetime.now()

files = [f for f in os.listdir('.') if os.path.isfile(f)]
for count,f in enumerate(files):
    xml_data = open(f, 'r').read()  # Read file
    root = et.XML(xml_data)  # Parse XML
    
    # Move resolução para diretório Resolucao
    if not 'Resolucao' in os.listdir(): os.mkdir("Resolucao")
    if root.find('TipoDocumento').text == "Resolução": 
        os.rename(os.getcwd()+"\\"+f,os.getcwd()+"\\Resolucao\\"+f)
    else:
        # Move acórdão para diretório Acordao\versãoPGD
        if not 'Acordao' in os.listdir(): os.mkdir("Acordao")
        if not root.find('VersaoPGD').text in os.listdir("Acordao"): 
            os.chdir('Acordao')
            os.mkdir(str(root.find('VersaoPGD').text))
            os.chdir('..')
        versao_pgd = root.find('VersaoPGD').text
        os.rename(os.getcwd()+"\\"+f,os.getcwd()+"\\Acordao\\"+versao_pgd+"\\"+f)

    decorrido = datetime.now() - inicio
    expectativa = decorrido / (count+1) * len(files)
    if decorrido > expectativa: 
        restante = 0 
    else: 
        restante = expectativa - decorrido
    
    print (count, str(decorrido)[:-7], str(expectativa)[:-7], str(restante)[:-7], f, root.find('TipoDocumento').text, root.find('VersaoPGD').text)


os.chdir('Acordao\\4.0.0')
os.getcwd()
