==> 4.Create_DataFrame.ipynb <==
import xml.etree.ElementTree as et
import pandas as pd
import numpy as np
import os
import re
from pprint import pprint
from datetime import datetime
import nltk
from nltk.corpus import PlaintextCorpusReader
import nltk.data
# nltk.download()

# os.chdir('d:\\Users\\Patri\\Documents\\1.TCC\\Notebooks\\WebScrapAcordaosPDF')

pasta = os.listdir()
for count, arquivo in enumerate(pasta) : print (count, arquivo)

os.getcwd()

# Extrai informações dos arquivos XML para Pandas Database - versão PGD 4.0.0
datatable = []

import os
from datetime import datetime

inicio = datetime.now()

files = [f for f in os.listdir('.') if os.path.isfile(f)]
for count, arquivo in enumerate(files):
    xml_data = open(arquivo, 'r').read()  # Read file
    root = et.XML(xml_data)  # Parse XML    
    if root.find('TipoDocumento').text != "Resolução":
        if root.find('VersaoPGD').text == "4.0.0":    
            cols = []
            data = []

            for i in (0,1,2,4,5,6,7,9):
                cols.append(root[i].tag)
                data.append(root[i].text)

            ementa = ""
            ementas = root.findall('Ementa')
            
            for linha in ementas:
                for q in range(len(linha)):
                    ementa += str(linha[q].text)

            cols.append('Ementa')
            data.append(ementa)

            acordao = ""
            acordaos = root.findall('Acordao')

            for linha in acordaos:
                for q in range(len(linha)):
                    acordao += str(linha[q].text)

            cols.append('Acordao')
            data.append(acordao)            

            relatorio = root.findall('Relatorio')
            cols.append('Relatorio')
            data.append(relatorio[0].text)
            
            voto = root.findall('Voto')
            cols.append('Voto')
            data.append(voto[0].text)
                        
            votovencedor = root.findall('VotoVencedor')
            cols.append('VotoVencedor')
            data.append(votovencedor[0].text)
                        
            declaracaovoto = root.findall('DeclaracaoVoto')
            cols.append('DeclaracaoVoto')
            data.append(declaracaovoto[0].text)
            
            datatable.append(data)

    decorrido = datetime.now() - inicio
    expectativa = decorrido / (count+1) * len(files)
    if decorrido > expectativa: 
        restante = 0 
    else: 
        restante = expectativa - decorrido
    print(count, str(decorrido)[:-7], str(expectativa)[:-7], str(restante)[:-7], arquivo, root.find('TipoDocumento').text, root.find('VersaoPGD').text)

df = pd.DataFrame.from_records(datatable, columns=cols)

# Extrai informações dos arquivos XML para Pandas Database - versão PGD 2.0.4 ou 2.0.1
datatable = []

import os
from datetime import datetime

inicio = datetime.now()

files = [f for f in os.listdir('.') if os.path.isfile(f)]
for count, arquivo in enumerate(files):
    xml_data = open(arquivo, 'r').read()  # Read file
    root = et.XML(xml_data)  # Parse XML    
    if root.find('TipoDocumento').text != "Resolução":
        if root.find('VersaoPGD').text == "2.0.4" or root.find('VersaoPGD').text == "2.0.1":    
            cols = []
            data = []

            processo = root.findall('Processo')
            cols.append('Processo')
            data.append(processo[0].text)

            decisao = root.findall('Decisao')
            cols.append('Decisao')
            data.append(decisao[0].text)

            datas = root.findall('Data')
            cols.append('Data')
            data.append(datas[0].text)

            ementa = ""
            ementas = root.findall('Ementa')
            
            for linha in ementas:
                for q in range(len(linha)):
                    ementa += str(linha[q].text)

            cols.append('Ementa')
            data.append(ementa)

            acordao = ""
            acordaos = root.findall('Acordao')

            for linha in acordaos:
                for q in range(len(linha)):
                    acordao += str(linha[q].text)

            cols.append('Acordao')
            data.append(acordao)            
            
            relatorio = root.findall('Relatorio')
            cols.append('Relatorio')
            data.append(relatorio[0].text)
            
            voto = root.findall('Voto')
            cols.append('Voto')
            data.append(voto[0].text)
                        
            votovencedor = root.findall('VotoVencedor')
            cols.append('VotoVencedor')
            data.append(votovencedor[0].text)
                        
            declaracaovoto = root.findall('DeclaracaoVoto')
            cols.append('DeclaracaoVoto')
            data.append(declaracaovoto[0].text)
            
            datatable.append(data)

    decorrido = datetime.now() - inicio
    expectativa = decorrido / (count+1) * len(files)
    if decorrido > expectativa: 
        restante = 0 
    else: 
        restante = expectativa - decorrido
    print(count, str(decorrido)[:-7], str(expectativa)[:-7], str(restante)[:-7], arquivo, root.find('TipoDocumento').text, root.find('VersaoPGD').text)

df = pd.DataFrame.from_records(datatable, columns=cols)

df

os.chdir('d:\\Users\\Patri\\Documents\\1.TCC\\Notebooks\\WebScrapAcordaosPDF')

pasta_qual='NLTK'

os.chdir(pasta_qual)

df.to_csv('Acordaosprecedentes400.csv', index=False)

os.listdir()
