==> 5.Tratamento_de_dados.ipynb <==
import pandas as pd
import numpy as np
import os
from pprint import pprint

os.chdir('d:\\Users\\Patri\\Documents\\1.TCC\\Notebooks\\WebScrapAcordaosPDF\\NLTK')

%%time
df = pd.read_csv('Acordaos2020.csv')

df.head()

df = df.drop(columns='Unnamed: 0')

df.info()

df.isnull().sum()

# Tratamento dos dados

# # Remove colunas desnecessárias para a presente análise.
# # EDIT: Colunas já foram removidas na extração do XML

# # Todas as linhas da coluna 'CPF Relator' estão vazias.
# df = df.drop(columns=['CPFRelator']) # vazio

# # Todas as linhas da coluna 'VersaoPGD' estão preenchidas com '4.0.0'.
# df = df.drop(columns=['VersaoPGD']) # 4.0.0

# # Todas as linhas da coluna 'TipoDocumento' estão preenchidas com 'Acórdão'.
# df = df.drop(columns=['TipoDocumento']) # Acórdão

# # Todas as linhas da coluna 'Orgao' estão preenchidas com 'CARF'.
# df = df.drop(columns=['Orgao']) # CARF

# Conversão das colunas de texto em string, elimina valores nulos
colunas = ('NomeRelator', 'Ementa', 'Acordao', 'Relatorio', 'Voto', 'VotoVencedor', 'DeclaracaoVoto')
for coluna in colunas: df[coluna] = df[coluna].astype('str')

# Substitui por '' as células vazias, já que são de textos opcionais, exceto 'NomeRelator', que pode ser obtido da coluna 'Voto'
df = df.fillna('')
df = df.replace(np.nan,'')

# Converte a coluna 'Data' para o formato datetime
df['Data'] = pd.to_datetime(df['Data'])

# Remove acórdãos de sessões realizadas em anos anteriores, em que as súmulas mais recente não estavam em vigor
df = df[~(df['Data'] < '2020-01-01')]

df.isnull().sum()

df = df.reset_index(drop = True)

df.info()

%%time
df.to_csv('Acordaos2020tratado.csv', index=False)
