==> 6b.PreProcess_BoolearSumula.ipynb <==
import pandas as pd
import numpy as np
import os
from collections import Counter
from pprint import pprint
from datetime import datetime
from IPython.display import display
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

os.chdir('d:\\Users\\Patri\\Documents\\1.TCC\\Notebooks\\WebScrapAcordaosPDF\\NLTK')

%%time
df = pd.read_csv('Acordaos2020sumulado.csv')

df.head()

df.isnull().sum()

# Remover as colunas de textos que não serão utilizadas para a predição
colunas = ['Ementa', 'Acordao', 'Voto', 'VotoVencedor', 'DeclaracaoVoto']
t_colunas = list('t_' + coluna for coluna in colunas)
remover_colunas = colunas + t_colunas
remover_colunas

for coluna in remover_colunas:
    df = df.drop(columns=coluna)

df

# Remover as colunas de textos que não serão utilizadas para a predição
remover_colunas = ['Processo', 'TipoRecurso', 'Turma', 'Data', 'Recorrente', 'Recorrida', 'NomeRelator', 't_Relatorio']

for coluna in remover_colunas:
    df = df.drop(columns=coluna)

df

# Conversão das colunas de texto em string, elimina valores nulos
coluna = ('Relatorio')
df[coluna] = df[coluna].astype('str')

df = df.fillna('')
df = df.replace(np.nan,'')

df.info()

# Totalizar as súmulas aplicadas por coluna
colunas=['Voto', 'Ementa', 'Relatorio', 'Acordao', 'VotoVencedor', 'DeclaracaoVoto']

sum_total = []
for count, coluna in enumerate(colunas): 
    sum_coluna = []
    for index, values in df['s_'+coluna].items():
        if values: 
            lista = values.strip('][').split(', ')
            for sumula in lista:
                 if len(sumula):
                    sum_coluna.append(int(sumula))
                    sum_total.append(int(sumula))
        sum_coluna = sorted(list(set(sum_coluna)))    
    print ("Coluna:", coluna, len(sum_coluna), sum_coluna)
sum_total = sorted(list(set(sum_total)))
print ("Total:", len(sum_total), sum_total)


# Criar as colunas booleanas para cada sumula(161) e coluna(6):
colunas=['Voto', 'Ementa', 'Relatorio', 'Acordao', 'VotoVencedor', 'DeclaracaoVoto']
s_colunas=['s_Voto', 's_Ementa', 's_Relatorio', 's_Acordao', 's_VotoVencedor', 's_DeclaracaoVoto']
z_colunas=['VO', 'EM', 'RE', 'AC', 'VV', 'DE']
# Formato de nome das colunas: S001AC ('S' + numero da sumula em 3 digitos + 2 digitos da coluna)
# Criar todas as colunas com valores False

# loop por coluna
for count, coluna in enumerate(colunas): 
    print ("Coluna:", coluna, count, s_colunas[count], z_colunas[count])

    # criar colunas para as booleanas das sumulas
    for sumula in range (1,162):
        df['S'+str(sumula).zfill(3)+z_colunas[count]] = 0

%%time

# Pega lista de número das súmulas de cada célula e liga (1) a coluna t_seção correspondente
colunas=['Voto', 'Ementa', 'Relatorio', 'Acordao', 'VotoVencedor', 'DeclaracaoVoto']
s_colunas=['s_Voto', 's_Ementa', 's_Relatorio', 's_Acordao', 's_VotoVencedor', 's_DeclaracaoVoto']
z_colunas=['VO', 'EM', 'RE', 'AC', 'VV', 'DE']
# s_colunas = []
# for coluna in colunas: s_colunas.append('s_'+coluna)

# Identifica as súmulas aplicáveis na colunas de seção do acórdão e seta como '1' a respectiva coluna booleana
# Aplicada em uma s_coluna (coluna com lista de súmulas aplicáveis)
# Busca localizar uma sumula específica

# loop por coluna de seção
for count, coluna in enumerate(colunas): 
    s_coluna = 's_' + coluna    

    # loop por linha (célula a céula)
    lista_sumulas = df[s_coluna].items() #retorna uma tupla dupla (index, list)
    for index, value in lista_sumulas:
        if len(value):         
            lista_sumula = value.strip('][').split(', ')
            # pega cada número encontrado e liga (1) a coluna respectiva
            for numero_sumula in lista_sumula:
                 if len(numero_sumula):
                    col_bool = 'S' + str(numero_sumula).zfill(3) + z_colunas[count]
                    print (coluna, index, col_bool, lista_sumula, numero_sumula)
                    df.at[index, col_bool] = 1

df

%%time
df.to_csv('Acordaos2020booleado.csv', index = False)
