==> 7.Pycaret_NLP_Classification.ipynb <==
!pip install pycaret --user

!pip install spacy
# pip3 install https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz

!python -m spacy download en_core_web_sm --user
!python -m textblob.download_corpora

!python -m spacy download pt_core_news_lg --user
!python -m textblob.download_corpora

import pandas as pd
import numpy as np
import spacy
import os
from sklearn.model_selection import train_test_split
from pycaret.nlp import *

spacy.load('en_core_web_sm')

spacy.load('pt_core_news_lg')

path = 'd:\\Users\\Patri\\Documents\\1.TCC\\Notebooks\\WebScrapAcordaosPDF\\NLTK'
os.chdir(path)

# Ler Arquivo gravado com todo o tratamento dado até aqui
dfcsv = pd.read_csv('Acordaos2020tx.csv')

sum_num = 157
sum_str = 'S'+str(sum_num).zfill(3)     #S999
sum_RE = sum_str + 'RE'
sum_VO = sum_str + 'VO'

dfcsv[[sum_RE, sum_VO]].value_counts() 

# selecionar aplicações da súmula como precedentes
dataset_precedentes = dfcsv[['Relatorio', sum_VO]].query(sum_VO + ' == 1')
dataset_precedentes

df_RE0 = dfcsv.query(sum_RE + ' == 0').query(sum_VO + ' == 0')   # Selecionar apenas acórdãos cujos relatórios e votos não aplicaram a súmula
df_RE0[[sum_RE, sum_VO]].value_counts()

df, validation = train_test_split(df_RE0, train_size = 2000)
print(len (df), len (validation))
df[[sum_VO]].value_counts()

df = df[['Relatorio', sum_VO]]
df = df.reset_index(drop = True)
dataset_2020 = df
dataset_2020

path = 'd:\\Users\\Patri\\Documents\\1.TCC\\Notebooks\\WebScrapAcordaosPDF\\NLTK'
# dataset_precedentes = pd.read_csv(path+'\\dataset_precedentes_' + str(sum_num) + '.csv')
dataset_precedentes

dataset_pycaret = dataset_2020

dataset_pycaret = pd.concat ([dataset_2020, dataset_precedentes], ignore_index=True)
dataset_pycaret

# dataset_pycaret.drop(['Decisao', 's_Relatorio'], axis=1, inplace = True)

# dataset_pycaret.query(sum_VO + ' == 1')
dataset_pycaret

ptstopwords = ['º', 'ª', 'a', 'ainda', 'ao', 'aos', 'apenas', 'apos', 'após', 'aquela', 'aquelas', 'aquele', 'aqueles', 'aquilo', 'art', 'artigo', 'arts', 'as', 'assim', 'assunto', 'até', 'brasil', 'cabe', 'caracteriza', 'caso', 'com', 'como', 'conforme', 'correspondente', 'da', 'das', 'data', 'de', 'dela', 'delas', 'dele', 'deles', 'depois', 'desde', 'direito', 'do', 'dos', 'e', 'ela', 'elas', 'ele', 'eles', 'em', 'entre', 'era', 'eram', 'essa', 'essas', 'esse', 'esses', 'esta', 'estamos', 'estas', 'estava', 'estavam', 'este', 'esteja', 'estejam', 'estejamos', 'estes', 'esteve', 'estive', 'estivemos', 'estiver', 'estivera', 'estiveram', 'estiverem', 'estivermos', 'estivesse', 'estivessem', 'estivéramos', 'estivéssemos', 'estou', 'está', 'estávamos', 'estão', 'eu', 'falar', 'fl', 'fls', 'foi', 'fomos', 'for', 'fora', 'foram', 'forem', 'formos', 'fosse', 'fossem', 'fui', 'fôramos', 'fôssemos', 'haja', 'hajam', 'hajamos', 'havemos', 'hei', 'houve', 'houvemos', 'houver', 'houvera', 'houveram', 'houverei', 'houverem', 'houveremos', 'houveria', 'houveriam', 'houvermos', 'houverá', 'houverão', 'houveríamos', 'houvesse', 'houvessem', 'houvéramos', 'houvéssemos', 'há', 'hão', 'ii', 'in', 'inciso', 'isso', 'isto', 'já', 'lhe', 'lhes', 'mais', 'mas', 'me', 'mesma', 'mesmo', 'meu', 'meus', 'minha', 'minhas', 'muito', 'n', 'na', 'nao', 'nas', 'nem', 'no', 'nos', 'nossa', 'nossas', 'n', 'nosso', 'nossos', 'num', 'numa', 'nº', 'não', 'nós', 'o', 'os', 'ou', 'para', 'pela', 'pelas', 'pelo', 'pelos', 'pode', 'pois', 'por', 'qual', 'qualquer', 'quando', 'quanto', 'que', 'quem', 'se', 'seja', 'sejam', 'sejamos', 'sem', 'sendo', 'ser', 'serei', 'seremos', 'seria', 'seriam', 'será', 'serão', 'seríamos', 'seu', 'seus', 'sido', 'sobre', 'somos', 'sou', 'sua', 'suas', 'são', 'só', 'tais', 'tal', 'tambem', 'também', 'te', 'tem', 'temos', 'tendo', 'tenha', 'tenham', 'tenhamos', 'tenho', 'ter', 'terei', 'teremos', 'teria', 'teriam', 'termos', 'terá', 'terão', 'teríamos', 'teu', 'teus', 'teve', 'tinha', 'tinham', 'tive', 'tivemos', 'tiver', 'tivera', 'tiveram', 'tiverem', 'tivermos', 'tivesse', 'tivessem', 'tivéramos', 'tivéssemos', 'todos', 'trata', 'tu', 'tua', 'tuas', 'tém', 'tínhamos', 'um', 'uma', 'valor', 'valores', 'vez', 'você', 'vocês', 'vos', 'à', 'às', 'é', 'éramos']

setup(data = dataset_pycaret, target = 'Relatorio', custom_stopwords=ptstopwords)
# setup(data = dataset_pycaret, target = 'Relatorio')

lda = create_model('lda', num_topics=4, multi_core = True)
lda_results = assign_model(lda)
lda_results.head()

plot_model(lda, plot = 'topic_distribution')

plot_model(lda, plot = 'frequency', topic_num = 'Topic 0')

plot_model(lda, plot = 'frequency', topic_num = 'Topic 1')

plot_model(plot = 'bigram')

plot_model(lda, plot = 'tsne')

plot_model(lda, plot = 'umap')

evaluate_model(lda)

lda_results

lda_results.drop(['Relatorio', 'Dominant_Topic', 'Perc_Dominant_Topic'], axis=1, inplace = True)

lda_results.head()

from pycaret.classification import *

clsfi = setup (data = lda_results, target = sum_VO, session_id = sum_num, train_size = 0.75)

compare_models()

rf = create_model('rf')
evaluate_model(rf)

tuned_rf = tune_model(rf, optimize='F1')
evaluate_model(tuned_rf)

boosted_rf = ensemble_model(rf, method = 'Boosting', n_estimators = 10)
evaluate_model(boosted_rf)

nb = create_model('nb')
evaluate_model(nb)

tuned_nb = tune_model(nb, optimize='Recall')
evaluate_model(tuned_nb)

boosted_nb = ensemble_model(nb, method = 'Boosting', n_estimators = 3)
evaluate_model(boosted_nb)

nb = create_model('nb')
boosted_nb = ensemble_model(nb, method = 'Boosting', n_estimators = 3)
evaluate_model(boosted_nb)
modelo_final = finalize_model(boosted_nb)
save_model (boosted_nb, sum_str)

data_unseen = validation[['Relatorio', sum_VO]]
predictions = predict_model(modelo_final, data = data_unseen)

nb = create_model('nb')
# evaluate_model(nb)

# tuned_nb = tune_model(nb, optimize='Recall')
# evaluate_model(tuned_nb)

ens_nb = ensemble_model(nb, method = 'Bagging')
evaluate_model(ens_nb)

# boosted_nb = ensemble_model(dt, method = 'Boosting', n_estimators = 100)
# evaluate_model(boosted_nb)



# lr = create_model('lr')
# tuned_lr=tune_model(lr)
boosted_lr = ensemble_model(lr, method = 'Boosting', n_estimators = 100)
evaluate_model(boosted_lr)

rf = create_model('rf')
tuned_rf = tune_model(rf)
evaluate_model(rf)

lda = create_model('lda')
tuned_lda=tune_model(lda, optimize='F1')
evaluate_model(tuned_lda)

predictions = predict_model(lda)

predictions.query("S137VO == '1'")

boost_dt= ensemble_model(dt, method = "Bagging", n_estimators=100)

evaluate_model(dt)

knn = create_model('knn')

evaluate_model(knn)

nb = create_model('nb')

nb_t = tune_model(nb, optimize='F1')

evaluate_model(nb)

gbc = create_model('gbc')

evaluate_model(gbc)

dt = create_model('dt')

evaluate_model(dt)

evaluate_model(lda)

save_model(lda,'Testing LDA 2020 e precedentes')
